move features module
